Add CreateExistsProcessConfigurationException and AlterNotExistProcessConfigurationException
