+++ docs/document/content/user-manual/error-code/sql-error-code.en.md
@@ -89,8 +89,10 @@ SQL error codes provide by standard `SQL State`, `Vendor Code` and `Reason`, whi
 
 | SQL State | Vendor Code | Reason |
 | --------- | ----------- | ------ |
-| HY000     | 18000       | Can not find pipeline job \`%s\` |
-| HY000     | 18001       | Failed to get DDL for table \`%s\` |
+| 44000     | 18001       | Created process configuration already existed |
+| 44000     | 18002       | Altered process configuration does not exist |
+| HY000     | 18020       | Failed to get DDL for table \`%s\` |
+| HY000     | 18080       | Can not find pipeline job \`%s\` |
 
 ## Feature Exception
